requirements.txt
Flask>=2.0.1
Flask-Cors>=3.0.10
google-generativeai
chromadb
pypdf
python-docx
textract==1.6.3
python-pptx
python-dotenv
sentence-transformers
